push to render.com

+++ db/migrate/20241204134515_remove_phone_from_users.rb
@@ -1,5 +1,7 @@
 class RemovePhoneFromUsers < ActiveRecord::Migration[6.1]
   def change
-    remove_column :users, :phone, :string
+    if column_exists?(:users, :phone)
+      remove_column :users, :phone, :string
+    end
   end
 end
